after defense

+++ database/migrations/2020_07_19_154452_create_chart_totalquantityandcropprice_view.php
@@ -19,7 +19,7 @@ class CreateChartTotalquantityandcroppriceView extends Migration
         SELECT
         posts.crop_name AS crop_name,
         SUM(posts.crop_quantity::decimal) AS sumcropqty,
-        (SUM(posts.crop_quantity::decimal) * posts.crop_price) AS allcropprice,
+        (SUM(posts.crop_quantity::decimal) * posts.crop_price::decimal) AS allcropprice,
         AVG(posts.crop_price::decimal) AS avgcropprice
         FROM
         posts
